requirements.txt
flask
openai
python-dotenv
flask-cors
python-docx
PyPDF2
markdown
langdetect
pandas
openpyxl
pytest
